api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
requests==2.31.0
python-dotenv==1.0.0
